Prepare/350.0.x

+++ nekoyume/Assets/_Scripts/UI/Model/ShopEquipments.cs
@@ -23,7 +23,7 @@ namespace Nekoyume.UI.Model
         public string sellerAgentAddress;
         public string sellerAvatarAddress;
         public decimal price;
-        public int combatPoint;
+        public long combatPoint;
         public int level;
         public int id;
         public int itemCount;
